rebuild/requirements.txt
flask
gunicorn
psycopg2-binary
sqlalchemy
flask-sqlalchemy
flask-login
pyswisseph
astral
ephem
pytz
pandas
tabulate
icalendar
requests
